+++ gDRplots/inst/scripts/PRISM_test_data.R
@@ -46,6 +46,7 @@ data.table::fwrite(tab_cko,
                    file = file.path(testdata_dir, "CRISPRGeneEffect.csv"), 
                    row.names = FALSE)
 
+
 # OmicsSomaticMutationsMatrixHotspot.csv ----
 tab_hot <- data.table::data.table(
   V1 = tab_model$ModelID,
@@ -60,6 +61,7 @@ tab_hot[[3]][11:12] <- NA
 data.table::fwrite(tab_hot, 
                    file = file.path(testdata_dir, "OmicsSomaticMutationsMatrixHotspot.csv"), 
                    row.names = FALSE)
+
 
 # OmicsSignaturesProfile ----
 chars <- c(letters, LETTERS, 0:9)
@@ -77,6 +79,25 @@ tab_profil <- data.table::data.table(
 data.table::fwrite(tab_profil, 
                    file = file.path(testdata_dir, "OmicsSignaturesProfile.csv"), 
                    row.names = FALSE)
+
+
+# OmicsArmLevelCNA ----
+tab_arm <- data.table::data.table(
+  V1 = tab_model$ModelID,
+  `1p` = withr::with_seed(42, sample(c(-1, 0, 1), size = NROW(tab_model), prob = c(0.20, 0.75, 0.05), replace = TRUE)),
+  `1q` = withr::with_seed(42, sample(c(-1, 0, 1), size = NROW(tab_model), prob = c(0.08, 0.8, 0.12), replace = TRUE)),
+  `8p` = withr::with_seed(314, sample(c(-1, 0, 1), size = NROW(tab_model), prob = c(0.20, 0.75, 0.05), replace = TRUE)),
+  `8q` = withr::with_seed(314, sample(c(-1, 0, 1), size = NROW(tab_model), prob = c(0.08, 0.8, 0.12), replace = TRUE)),
+  `16p` = withr::with_seed(42, sample(c(-1, 0, 1), size = NROW(tab_model), prob = c(0.15, 0.70, 0.15), replace = TRUE)),
+  `16q` = withr::with_seed(42, sample(c(-1, 0, 1), size = NROW(tab_model), prob = c(0.09, 0.7, 0.21), replace = TRUE)),
+  `22p` = withr::with_seed(314, sample(c(-1, 0, 1), size = NROW(tab_model), prob = c(0.15, 0.70, 0.15), replace = TRUE)),
+  `22q` = withr::with_seed(314, sample(c(-1, 0, 1), size = NROW(tab_model), prob = c(0.09, 0.7, 0.21), replace = TRUE))
+)
+
+data.table::fwrite(tab_arm, 
+                   file = file.path(testdata_dir, "OmicsArmLevelCNA.csv"), 
+                   row.names = FALSE)
+
 
 # assoc table ----
 tab_drug_001_met1_RV <- data.table::data.table(